Added bin details

+++ routes/userpoints.js
@@ -13,23 +13,28 @@ router.get("/updateUserPoints", (req, res) => {
     const trasherId = req.query.trasherId;
     const binId = req.query.binId;
     BinWeights.find({ binId: binId }, (err, docs) => {
-        const weight = docs[0].weight;
-        var points = weight
-        UserPoints.find({ trasherId: trasherId }, (err, pointResult) => {
-            const currentUserPoints = pointResult[0].points;
-            const updatedPoints = currentUserPoints + points
-            UserPoints.updateOne({ trasherId: trasherId }, { $set: { points: updatedPoints } }, (err, result) => {
-                if (!err) {
-                    BinWeights.deleteOne({ binId: binId }, (err, binDeleteResult) => {
-                        res.json({ message: "Updated", points: updatedPoints })
-                    })
+        if (docs.length !== 0) {
+            const weight = docs[0].weight;
+            var points = weight
+            UserPoints.find({ trasherId: trasherId }, (err, pointResult) => {
+                console.log(pointResult)
+                const currentUserPoints = pointResult[0].points;
+                const updatedPoints = currentUserPoints + points
+                UserPoints.updateOne({ trasherId: trasherId }, { $set: { points: updatedPoints } }, (err, result) => {
+                    if (!err) {
+                        BinWeights.deleteOne({ binId: binId }, (err, binDeleteResult) => {
+                            res.json({ message: "Updated", points: updatedPoints })
+                        })
 
-                }
-                else {
-                    res.json({ message: err })
-                }
+                    }
+                    else {
+                        res.json({ message: err })
+                    }
+                })
             })
-        })
+        } else {
+            res.json({ message: 'Token Expired!' })
+        }
     })
 })
